requirements.txt
aiohttp
aiofiles
aiosqlite
asyncio
dateparser
oauth2client
google-api-python-client
yt-dlp[default]
curl-cffi
yp_dl
internetarchive
playwright
https://codeload.github.com/bindestriche/srt_fix/zip/refs/heads/master
Pillow>=9.0.0
streamlit>=1.28.0
plotly>=5.17.0
streamlit-community-navigation-bar
multipage_streamlit
streamlit-option-menu
streamlit_dynamic_filters
streamlit-aggrid
streamlit-tree-select
streamlit-echarts